requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==25.1.0
mysqlclient==2.2.8
openpyxl==3.1.5
packaging==26.0
PyJWT==2.11.0
python-decouple==3.8
sqlparse==0.5.5
